Bug fix and further refactoring module related stuff
