+++ app/app.go
@@ -9,6 +9,7 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/cosmos/cosmos-sdk/store/streaming"
 	"github.com/gorilla/mux"
 	"github.com/spf13/cast"
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -32,52 +33,16 @@ import (
 	"github.com/cosmos/cosmos-sdk/server/config"
 	servertypes "github.com/cosmos/cosmos-sdk/server/types"
 	"github.com/cosmos/cosmos-sdk/simapp"
-	"github.com/cosmos/cosmos-sdk/store/streaming"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	"github.com/cosmos/cosmos-sdk/version"
-	"github.com/cosmos/cosmos-sdk/x/auth"
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/cosmos/cosmos-sdk/x/auth/vesting"
-	vestingtypes "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
-	"github.com/cosmos/cosmos-sdk/x/authz"
-	authzmodule "github.com/cosmos/cosmos-sdk/x/authz/module"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/cosmos/cosmos-sdk/x/capability"
 	capabilitykeeper "github.com/cosmos/cosmos-sdk/x/capability/keeper"
-	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
 	"github.com/cosmos/cosmos-sdk/x/crisis"
-	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
-	distr "github.com/cosmos/cosmos-sdk/x/distribution"
-	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-	"github.com/cosmos/cosmos-sdk/x/evidence"
-	evidencetypes "github.com/cosmos/cosmos-sdk/x/evidence/types"
-	"github.com/cosmos/cosmos-sdk/x/feegrant"
-	feegrantmodule "github.com/cosmos/cosmos-sdk/x/feegrant/module"
-	"github.com/cosmos/cosmos-sdk/x/genutil"
-	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-	"github.com/cosmos/cosmos-sdk/x/gov"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	"github.com/cosmos/cosmos-sdk/x/mint"
-	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
-	"github.com/cosmos/cosmos-sdk/x/params"
-	paramstypes "github.com/cosmos/cosmos-sdk/x/params/types"
-	"github.com/cosmos/cosmos-sdk/x/slashing"
-	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/cosmos/cosmos-sdk/x/upgrade"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 
 	"github.com/dymensionxyz/dymension/v3/docs"
 
-	ibctransfer "github.com/cosmos/ibc-go/v6/modules/apps/transfer"
-	ibctransfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
-	ibc "github.com/cosmos/ibc-go/v6/modules/core"
-	ibchost "github.com/cosmos/ibc-go/v6/modules/core/24-host"
 	ibckeeper "github.com/cosmos/ibc-go/v6/modules/core/keeper"
 	ibctesting "github.com/cosmos/ibc-go/v6/testing"
 	ibctestingtypes "github.com/cosmos/ibc-go/v6/testing/types"
@@ -86,24 +51,6 @@ import (
 
 	"github.com/dymensionxyz/dymension/v3/app/ante"
 	appparams "github.com/dymensionxyz/dymension/v3/app/params"
-
-	rollappmodule "github.com/dymensionxyz/dymension/v3/x/rollapp"
-	rollappmoduletypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
-
-	sequencermodule "github.com/dymensionxyz/dymension/v3/x/sequencer"
-	sequencermoduletypes "github.com/dymensionxyz/dymension/v3/x/sequencer/types"
-
-	streamermodule "github.com/dymensionxyz/dymension/v3/x/streamer"
-	streamermoduletypes "github.com/dymensionxyz/dymension/v3/x/streamer/types"
-
-	denommetadatamodule "github.com/dymensionxyz/dymension/v3/x/denommetadata"
-	denommetadatamoduletypes "github.com/dymensionxyz/dymension/v3/x/denommetadata/types"
-
-	delayedackmodule "github.com/dymensionxyz/dymension/v3/x/delayedack"
-	delayedacktypes "github.com/dymensionxyz/dymension/v3/x/delayedack/types"
-
-	eibcmodule "github.com/dymensionxyz/dymension/v3/x/eibc"
-	eibcmoduletypes "github.com/dymensionxyz/dymension/v3/x/eibc/types"
 
 	packetforwardmiddleware "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v6/packetforward"
 	packetforwardkeeper "github.com/cosmos/ibc-apps/middleware/packet-forward-middleware/v6/packetforward/keeper"
@@ -115,28 +62,7 @@ import (
 
 	"github.com/evmos/ethermint/server/flags"
 	ethermint "github.com/evmos/ethermint/types"
-	"github.com/evmos/ethermint/x/evm"
-	evmtypes "github.com/evmos/ethermint/x/evm/types"
-	"github.com/evmos/ethermint/x/feemarket"
-	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
-
-	/* ----------------------------- osmosis imports ---------------------------- */
-
-	"github.com/osmosis-labs/osmosis/v15/x/epochs"
-	epochstypes "github.com/osmosis-labs/osmosis/v15/x/epochs/types"
-	"github.com/osmosis-labs/osmosis/v15/x/lockup"
-	lockuptypes "github.com/osmosis-labs/osmosis/v15/x/lockup/types"
-
-	"github.com/osmosis-labs/osmosis/v15/x/gamm"
-	gammtypes "github.com/osmosis-labs/osmosis/v15/x/gamm/types"
-	"github.com/osmosis-labs/osmosis/v15/x/incentives"
-	incentivestypes "github.com/osmosis-labs/osmosis/v15/x/incentives/types"
-	"github.com/osmosis-labs/osmosis/v15/x/poolmanager"
-	poolmanagertypes "github.com/osmosis-labs/osmosis/v15/x/poolmanager/types"
-
-	"github.com/osmosis-labs/osmosis/v15/x/txfees"
-	txfeestypes "github.com/osmosis-labs/osmosis/v15/x/txfees/types"
-	/* ---------------------------- upgrade handlers ---------------------------- */)
+	/* ----------------------------- osmosis imports ---------------------------- */ /* ---------------------------- upgrade handlers ---------------------------- */)
 
 var (
 	_ = packetforwardkeeper.DefaultForwardTransferPacketTimeoutTimestamp
@@ -153,26 +79,6 @@ var (
 	// ModuleBasics defines the module BasicManager is in charge of setting up basic,
 	// non-dependant module elements, such as codec registration
 	// and genesis verification.
-
-	// module account permissions
-	maccPerms = map[string][]string{
-		authtypes.FeeCollectorName:                         nil,
-		distrtypes.ModuleName:                              nil,
-		minttypes.ModuleName:                               {authtypes.Minter},
-		stakingtypes.BondedPoolName:                        {authtypes.Burner, authtypes.Staking},
-		stakingtypes.NotBondedPoolName:                     {authtypes.Burner, authtypes.Staking},
-		govtypes.ModuleName:                                {authtypes.Burner},
-		ibctransfertypes.ModuleName:                        {authtypes.Minter, authtypes.Burner},
-		sequencermoduletypes.ModuleName:                    {authtypes.Minter, authtypes.Burner, authtypes.Staking},
-		rollappmoduletypes.ModuleName:                      {},
-		streamermoduletypes.ModuleName:                     nil,
-		evmtypes.ModuleName:                                {authtypes.Minter, authtypes.Burner}, // used for secure addition and subtraction of balance using module account.
-		evmtypes.ModuleVirtualFrontierContractDeployerName: nil,                                  // used for deploying virtual frontier bank contract.
-		gammtypes.ModuleName:                               {authtypes.Minter, authtypes.Burner},
-		lockuptypes.ModuleName:                             {authtypes.Minter, authtypes.Burner},
-		incentivestypes.ModuleName:                         {authtypes.Minter, authtypes.Burner},
-		txfeestypes.ModuleName:                             {authtypes.Burner},
-	}
 )
 
 var (
@@ -248,22 +154,22 @@ func New(
 		appCodec:          appCodec,
 		interfaceRegistry: interfaceRegistry,
 		invCheckPeriod:    invCheckPeriod,
+		AppKeepers:        keepers.AppKeepers{},
 	}
 
-	appKeepers := keepers.AppKeepers{}
-	tracer := cast.ToString(appOpts.Get(flags.EVMTracer))
-
-	appKeepers.InitSpecialKeepers(appCodec, bApp, cdc, invCheckPeriod, skipUpgradeHeights, homePath)
-	appKeepers.InitNormalKeepers(appCodec, bApp, maccPerms, app.ModuleAccountAddrs(), tracer)
-	appKeepers.SetupHooks()
-	appKeepers.InitTransferStack()
-
-	app.AppKeepers = appKeepers
+	app.GenerateKeys()
 
 	// load state streaming if enabled
 	if _, _, err := streaming.LoadStreamingServices(bApp, appOpts, appCodec, keepers.KVStoreKeys); err != nil {
 		panic("failed to load state streaming services: " + err.Error())
 	}
+
+	tracer := cast.ToString(appOpts.Get(flags.EVMTracer))
+
+	app.AppKeepers.InitSpecialKeepers(appCodec, bApp, cdc, invCheckPeriod, skipUpgradeHeights, homePath)
+	app.AppKeepers.InitNormalKeepers(appCodec, bApp, app.ModuleAccountAddrs(), tracer)
+	app.AppKeepers.SetupHooks()
+	app.AppKeepers.InitTransferStack()
 
 	/****  Module Options ****/
 
@@ -273,171 +179,22 @@ func New(
 
 	// NOTE: Any module instantiated in the module manager that is later modified
 	// must be passed by reference here.
-	app.mm = module.NewManager(
-		genutil.NewAppModule(
-			app.AccountKeeper, app.StakingKeeper, app.BaseApp.DeliverTx,
-			encodingConfig.TxConfig,
-		),
-		auth.NewAppModule(appCodec, *app.AccountKeeper, nil),
-		authzmodule.NewAppModule(appCodec, *app.AuthzKeeper, app.AccountKeeper, app.BankKeeper, app.interfaceRegistry),
-		vesting.NewAppModule(*app.AccountKeeper, app.BankKeeper),
-		bank.NewAppModule(appCodec, app.BankKeeper, app.AccountKeeper),
-		capability.NewAppModule(appCodec, *app.CapabilityKeeper),
-		feegrantmodule.NewAppModule(appCodec, app.AccountKeeper, app.BankKeeper, *app.FeeGrantKeeper, app.interfaceRegistry),
-		crisis.NewAppModule(app.CrisisKeeper, skipGenesisInvariants),
-		gov.NewAppModule(appCodec, *app.GovKeeper, app.AccountKeeper, app.BankKeeper),
-		mint.NewAppModule(appCodec, *app.MintKeeper, app.AccountKeeper, nil),
-		slashing.NewAppModule(appCodec, *app.SlashingKeeper, app.AccountKeeper, app.BankKeeper, app.StakingKeeper),
-		distr.NewAppModule(appCodec, *app.DistrKeeper, app.AccountKeeper, app.BankKeeper, app.StakingKeeper),
-		staking.NewAppModule(appCodec, *app.StakingKeeper, app.AccountKeeper, app.BankKeeper),
-		upgrade.NewAppModule(*app.UpgradeKeeper),
-		evidence.NewAppModule(*app.EvidenceKeeper),
-		ibc.NewAppModule(app.IBCKeeper),
-		params.NewAppModule(*app.ParamsKeeper),
-		packetforwardmiddleware.NewAppModule(app.PacketForwardMiddlewareKeeper),
-		ibctransfer.NewAppModule(*app.TransferKeeper),
-		rollappmodule.NewAppModule(appCodec, app.RollappKeeper, app.AccountKeeper, app.BankKeeper),
-		sequencermodule.NewAppModule(appCodec, *app.SequencerKeeper, app.AccountKeeper, app.BankKeeper),
-		streamermodule.NewAppModule(*app.StreamerKeeper, app.AccountKeeper, app.BankKeeper, app.EpochsKeeper),
-		delayedackmodule.NewAppModule(appCodec, *app.DelayedAckKeeper),
-		denommetadatamodule.NewAppModule(app.DenomMetadataKeeper, *app.EvmKeeper, app.BankKeeper),
-		eibcmodule.NewAppModule(appCodec, *app.EIBCKeeper, app.AccountKeeper, app.BankKeeper),
-		// this line is used by starport scaffolding # stargate/app/appModule
-
-		// Ethermint app modules
-		evm.NewAppModule(app.EvmKeeper, app.AccountKeeper, app.BankKeeper, app.GetSubspace(evmtypes.ModuleName).WithKeyTable(evmtypes.ParamKeyTable())),
-		feemarket.NewAppModule(*app.FeeMarketKeeper, app.GetSubspace(feemarkettypes.ModuleName).WithKeyTable(feemarkettypes.ParamKeyTable())),
-
-		// osmosis modules
-		lockup.NewAppModule(*app.LockupKeeper, app.AccountKeeper, app.BankKeeper),
-		epochs.NewAppModule(*app.EpochsKeeper),
-		gamm.NewAppModule(appCodec, *app.GAMMKeeper, app.AccountKeeper, app.BankKeeper),
-		poolmanager.NewAppModule(*app.PoolManagerKeeper, app.GAMMKeeper),
-		incentives.NewAppModule(*app.IncentivesKeeper, app.AccountKeeper, app.BankKeeper, app.EpochsKeeper),
-		txfees.NewAppModule(*app.TxFeesKeeper),
-	)
+	app.mm = module.NewManager(app.GetModules(appCodec, bApp, encodingConfig, skipGenesisInvariants)...)
 
 	// During begin block slashing happens after distr.BeginBlocker so that
 	// there is nothing left over in the validator fee pool, so as to keep the
 	// CanWithdrawInvariant invariant.
 	// NOTE: staking module is required if HistoricalEntries param > 0
 	// TODO: use "github.com/osmosis-labs/osmosis/osmoutils/partialord" to order modules
-	app.mm.SetOrderBeginBlockers(
-		epochstypes.ModuleName,
-		upgradetypes.ModuleName,
-		capabilitytypes.ModuleName,
-		minttypes.ModuleName,
-		distrtypes.ModuleName,
-		slashingtypes.ModuleName,
-		evidencetypes.ModuleName,
-		stakingtypes.ModuleName,
-		vestingtypes.ModuleName,
-		feemarkettypes.ModuleName,
-		evmtypes.ModuleName,
-		ibchost.ModuleName,
-		ibctransfertypes.ModuleName,
-		packetforwardtypes.ModuleName,
-		authtypes.ModuleName,
-		authz.ModuleName,
-		banktypes.ModuleName,
-		govtypes.ModuleName,
-		crisistypes.ModuleName,
-		genutiltypes.ModuleName,
-		feegrant.ModuleName,
-		paramstypes.ModuleName,
-		rollappmoduletypes.ModuleName,
-		sequencermoduletypes.ModuleName,
-		streamermoduletypes.ModuleName,
-		denommetadatamoduletypes.ModuleName,
-		delayedacktypes.ModuleName,
-		eibcmoduletypes.ModuleName,
-		// this line is used by starport scaffolding # stargate/app/beginBlockers
-		lockuptypes.ModuleName,
-		gammtypes.ModuleName,
-		poolmanagertypes.ModuleName,
-		incentivestypes.ModuleName,
-		txfeestypes.ModuleName,
-	)
-
-	app.mm.SetOrderEndBlockers(
-		crisistypes.ModuleName,
-		govtypes.ModuleName,
-		stakingtypes.ModuleName,
-		capabilitytypes.ModuleName,
-		authtypes.ModuleName,
-		authz.ModuleName,
-		banktypes.ModuleName,
-		distrtypes.ModuleName,
-		feemarkettypes.ModuleName,
-		evmtypes.ModuleName,
-		slashingtypes.ModuleName,
-		vestingtypes.ModuleName,
-		minttypes.ModuleName,
-		genutiltypes.ModuleName,
-		evidencetypes.ModuleName,
-		feegrant.ModuleName,
-		paramstypes.ModuleName,
-		upgradetypes.ModuleName,
-		ibchost.ModuleName,
-		ibctransfertypes.ModuleName,
-		packetforwardtypes.ModuleName,
-		rollappmoduletypes.ModuleName,
-		sequencermoduletypes.ModuleName,
-		streamermoduletypes.ModuleName,
-		denommetadatamoduletypes.ModuleName,
-		delayedacktypes.ModuleName,
-		eibcmoduletypes.ModuleName,
-		// this line is used by starport scaffolding # stargate/app/endBlockers
-		epochstypes.ModuleName,
-		lockuptypes.ModuleName,
-		gammtypes.ModuleName,
-		poolmanagertypes.ModuleName,
-		incentivestypes.ModuleName,
-		txfeestypes.ModuleName,
-	)
+	app.mm.SetOrderBeginBlockers(keepers.BeginBlockers...)
+	app.mm.SetOrderEndBlockers(keepers.EndBlockers...)
 
 	// NOTE: The genutils module must occur after staking so that pools are
 	// properly initialized with tokens from genesis accounts.
 	// NOTE: Capability module must occur first so that it can initialize any capabilities
 	// so that other modules that want to create or claim capabilities afterwards in InitChain
 	// can do so safely.
-	app.mm.SetOrderInitGenesis(
-		capabilitytypes.ModuleName,
-		authtypes.ModuleName,
-		authz.ModuleName,
-		banktypes.ModuleName,
-		distrtypes.ModuleName,
-		stakingtypes.ModuleName,
-		vestingtypes.ModuleName,
-		slashingtypes.ModuleName,
-		feemarkettypes.ModuleName,
-		evmtypes.ModuleName,
-		govtypes.ModuleName,
-		minttypes.ModuleName,
-		crisistypes.ModuleName,
-		ibchost.ModuleName,
-		genutiltypes.ModuleName,
-		evidencetypes.ModuleName,
-		paramstypes.ModuleName,
-		upgradetypes.ModuleName,
-		ibctransfertypes.ModuleName,
-		packetforwardtypes.ModuleName,
-		feegrant.ModuleName,
-		rollappmoduletypes.ModuleName,
-		sequencermoduletypes.ModuleName,
-		streamermoduletypes.ModuleName,
-		denommetadatamoduletypes.ModuleName, // must after `x/bank` to trigger hooks
-		delayedacktypes.ModuleName,
-		eibcmoduletypes.ModuleName,
-		// this line is used by starport scaffolding # stargate/app/initGenesis
-
-		epochstypes.ModuleName,
-		lockuptypes.ModuleName,
-		gammtypes.ModuleName,
-		poolmanagertypes.ModuleName,
-		incentivestypes.ModuleName,
-		txfeestypes.ModuleName,
-	)
+	app.mm.SetOrderInitGenesis(keepers.InitGenesis...)
 
 	app.mm.RegisterInvariants(app.CrisisKeeper)
 	app.mm.RegisterRoutes(app.Router(), app.QueryRouter(), encodingConfig.Amino)
@@ -515,19 +272,6 @@ func (app *App) LoadHeight(height int64) error {
 	return app.LoadVersion(height)
 }
 
-// ModuleAccountAddrs returns all the app's module account addresses.
-func (app *App) ModuleAccountAddrs() map[string]bool {
-	modAccAddrs := make(map[string]bool)
-	for acc := range maccPerms {
-		modAccAddrs[authtypes.NewModuleAddress(acc).String()] = true
-	}
-
-	// exclude the streamer as we want him to be able to get external incentives
-	modAccAddrs[authtypes.NewModuleAddress(streamermoduletypes.ModuleName).String()] = false
-	modAccAddrs[authtypes.NewModuleAddress(txfeestypes.ModuleName).String()] = false
-	return modAccAddrs
-}
-
 // LegacyAmino returns SimApp's amino codec.
 //
 // NOTE: This is solely to be used for testing purposes as it may be desirable
@@ -600,15 +344,6 @@ func RegisterSwaggerAPI(_ client.Context, rtr *mux.Router) {
 
 	staticServer := http.FileServer(http.FS(staticFS))
 	rtr.PathPrefix("/static/").Handler(http.StripPrefix("/static/", staticServer))
-}
-
-// GetMaccPerms returns a copy of the module account permissions
-func GetMaccPerms() map[string][]string {
-	dupMaccPerms := make(map[string][]string)
-	for k, v := range maccPerms {
-		dupMaccPerms[k] = v
-	}
-	return dupMaccPerms
 }
 
 // SimulationManager implements the SimulationApp interface
