+++ x/delayedack/keeper/params.go
@@ -7,14 +7,18 @@ import (
 )
 
 // GetParams get all parameters as types.Params
-func (k Keeper) GetParams(ctx sdk.Context) (params types.Params) {
-	k.paramstore.GetParamSet(ctx, &params)
-	return
+func (k Keeper) GetParams(ctx sdk.Context) types.Params {
+	return types.NewParams(k.EpochIdentifier(ctx), k.BridgingFee(ctx), k.DeletePacketBatchSize(ctx))
 }
 
 // SetParams set the params
 func (k Keeper) SetParams(ctx sdk.Context, params types.Params) {
 	k.paramstore.SetParamSet(ctx, &params)
+}
+
+func (k Keeper) EpochIdentifier(ctx sdk.Context) (res string) {
+	k.paramstore.Get(ctx, types.KeyEpochIdentifier, &res)
+	return
 }
 
 func (k Keeper) BridgingFee(ctx sdk.Context) (res sdk.Dec) {
@@ -24,4 +28,9 @@ func (k Keeper) BridgingFee(ctx sdk.Context) (res sdk.Dec) {
 
 func (k Keeper) BridgingFeeFromAmt(ctx sdk.Context, amt sdk.Int) (res sdk.Int) {
 	return k.BridgingFee(ctx).MulInt(amt).TruncateInt()
+}
+
+func (k Keeper) DeletePacketBatchSize(ctx sdk.Context) (res int32) {
+	k.paramstore.Get(ctx, types.KeyDeletePacketBatchSize, &res)
+	return
 }
